Update docs to make first example work as expected
Fixes https://github.com/chainlist/svelte-forms/issues/95 

It's expected that a form would behave as if it has checkOnInit on: empty form is invalid - false, once you add some input, it's true. 
One can of course read the documentation in detail to figure out what's wrong, but it might be useful for the first example to be self-sufficient.

The change was proposed by grischaerbe in the issue, and it is already explained in documentation. Thoughts on adding this to documentation?